Move coordinator into its own file & adjust code design more towards HA standards (WIP)
